+++ HideSeek/Common/Model/Reward.swift
@@ -10,12 +10,12 @@ class Reward {
     var pkId: Int64
     var name: String
     var imageUrl: String?
-    var record: Double = 0
+    var record: Int = 0
     var exchangeCount: Int
     var introduction: String?
     var version: Int64 = 0
     
-    init(pkId: Int64, name: String, imageUrl: String?, record: Double, exchangeCount: Int,
+    init(pkId: Int64, name: String, imageUrl: String?, record: Int, exchangeCount: Int,
          introduction: String?, version: Int64) {
         self.pkId = pkId
         self.name = name
