+++ HideSeek/DataAccess/ExchangeOrderTableManager.swift
@@ -19,7 +19,7 @@ class ExchangeOrderTableManager {
     let rewardId = Expression<Int64>("reward_id")
     let rewardName = Expression<String>("reward_name")
     let imageUrl = Expression<String>("image_url")
-    let record = Expression<Double>("record")
+    let record = Expression<Int>("record")
     let exchangeCount = Expression<Int>("exchange_count")
     let introduction = Expression<String>("introduction")
     let pullVersion = Expression<Int64>("version")
